Add lifespan management
