+++ src/bboard/main.py
@@ -8,9 +8,6 @@ Routes are defined here, with the actual logic appearing in neighboring modules.
 usage:  $ fastapi dev src/bboard/main.py
 """
 
-import asyncio
-from contextlib import asynccontextmanager
-
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 
@@ -18,19 +15,11 @@ from bboard.database import engine
 from bboard.demo.clock_display import clock_display, clock_reading
 from bboard.demo.greeting import greeting
 from bboard.models.iss_position import Base, IssPosition
-from bboard.transit.iss import iss_periodic_update, iss_world_map
-from bboard.transit.vehicles import query_vehicles, transit_periodic_update
+from bboard.transit.iss import iss_world_map
+from bboard.transit.vehicles import query_vehicles
+from bboard.util.lifespan_mgmt import lifespan
 from bboard.util.requests import patch_requests_module
 from bboard.util.web import table_of_contents
-
-
-@asynccontextmanager
-async def lifespan(app1: FastAPI) -> None:
-    assert app1
-    asyncio.create_task(iss_periodic_update())
-    asyncio.create_task(transit_periodic_update())
-    yield
-
 
 app = FastAPI(lifespan=lifespan)
